# Conflicts:
#   Plog_Backend/src/main/java/com/ssafy/plog/service/PostServiceImpl.java
#   plog_frontend/src/views/account/Profile.vue
#   plog_frontend/src/views/group/GroupNoteCreate.vue

+++ plog_frontend/src/views/group/GroupNoteDetail.vue
@@ -228,7 +228,12 @@ export default {
         getFormatDate(regtime) {
             return moment(new Date(regtime)).format('YYYY.MM.DD');
         },
-        deleteNote() {
+        deleteNote: async function () {
+          const res = await this.$dialog.warning({
+            text:"노트를 삭제 하시겠습니까?",
+            title: 'Delete Note'
+          });
+          if(res){
             http.delete('/post/', {
                 params : {
                   pId : this.pId,
@@ -271,6 +276,7 @@ export default {
                   this.$router.push("/404");
                 }                          
               });
+          }
         },
         makePDF () {
             var fileName = this.Note.pTitle;
